chara: Draw signed meshIndex<=0x1F compare (90.26->90.44)
Cast meshIndex to int so the bound check emits a signed cmpwi 0x20 matching
the target (Ghidra: 0x1f < (int)uVar8).

+++ src/chara.cpp
@@ -2159,7 +2159,7 @@ void CChara::CModel::Draw(float (*view)[4], int flags, int pass)
 		if (mesh->m_workPositions == 0) {
 			continue;
 		}
-		if (meshIndex <= 0x1F && ((ModelMeshVisibleMask(this) >> meshIndex) & 1) == 0) {
+		if (static_cast<int>(meshIndex) <= 0x1F && ((ModelMeshVisibleMask(this) >> meshIndex) & 1) == 0) {
 			continue;
 		}
